Fix deleting comments, do not allow blocked users to post comments.
Now, comments which have no children are deleted permanently, and
comments with replies are only marked as deleted. Also, a comment
which had replies won't be revived if you undelete its article,
and it may not be restored in its place after exporting and importing.
This depends on the order of articles inside the dump file, because
the parent comment must always exist when you are adding a new one.

Also add foreign key for wlc_parent->wlc_id and change foreign key for wlc_comment_page to SET NULL.